Refactor: 홈페이지 마켓 게시글 컴포넌트 적용
홈페이지 마켓 게시글 컴포넌트 적용
