+++ src/components/button/iconBtn.jsx
@@ -64,6 +64,7 @@ export const MarketPostMoreBtn  = ({productId}) => {
   
   const onClick = () => {
     setIsOpen(true);
+    console.log(productId)
   };
 
   const onClickClose = (value) => {
